Support POSIX message queue
A file descriptor is used for implementing a POSIX message queue on
Linux. lsof reports the file descriptor as a REGular file.

This change introduce "PSXMQ" as a new file type for the
file descriptor.

Currently implementation requires mqueue file system must be
mounted to resolve the file type. It is mounted at /dev/mqueue
by default on Fedora. It seems that the file system is not mounted
by default on Ubuntu.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>