[linux] make struct uxsin usable even when HASEPTOPTS it not defined
struct uxsin was used before +|-E option was introduced.
For an example, check_unix() uses struct uxsin, and it existed even in
lsof-4.78C.

To implement +|-E option for unix sockets, some fields were added to
struct uxsin. Therefore the fields should be available only when
HASEPTOPTS is defined. However, the structure itself should be
available always in the linux dialect.

No update for 00DIST because this change is invisble to users.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>