Handle #145
For handle `Compile and install the latest git master version of lsof on Ubuntu 20.04. #145`

While contrasting to other manpages, such as iptables.8.gz, arp.8.gz and so on, i figure out `.so ./version` is not necessary.
Therefore,  delete `.so ./version` in Lsof.8 may be the best way to solve this problem, whichi i have tested successful in ubuntu 20.04.