fix hyphen manpage errors
This is another patch that Debian has been carrying for a while. The original
patch description, by Raoul:
	a lot of minus sign occurences in the lsof man page are missing the
	'\' escape sign and would therefore be interpreted as hyphens. This
	patch adds the escape signs where appropriate.

The reasoning for this is long and complex (as described here, for example:
https://www.spinics.net/lists/linux-man/msg18858.html), but a much simpler
explanation is here: https://www.caichinger.com/blog/2006/09/06/man_hyphens/

"Use '-' whenever you want a hyphen and '\-' when you want a minus sign."
[except in the name section use \-, and use \- for program options/switches.]

Groff outputs different things for - and \-, and someone can get unlucky
when cutting & pasting from the man page. This patch (submitted by Andres)
ensures that in every place we're describing an option/switch to lsof, we
escape the hyphen.

Author: Raoul Gunnar Borenius <borenius@dfn.de>
Reviewed-by: Nicholas Bamber <nicholas@periapt.co.uk>
Reviewed-by: Andres Salomon <dilinger@debian.org>