[FreeBSD] sys/files.h no longer needs _KERNEL defined to include it
sys/files.h exports a userland interface. It no longer requires _KERNEL
to be defined to access it, and hasn't since approximaely 2002 or
so. Remove all the cruft we used to allow us to include with _KERNEL.
We also no longer need to probe sys/systm.h to see if it has pause or
not, since it's not drawn in.