[netbsd] Import patch-ah from pkgsrc
----------------------------
revision 1.5
date: 2011-03-21 08:22:02 +0800;  author: jym;  state: Exp;  lines: +2 -2;
When we moved from fd_ofiles to a struct fdtab, the fd_ofiles became
an array of pointer (in struct fdtab) rather than a pointer of pointers.

Sadly for us, no, arrays and pointers are not equivalent from a memory
perspective: while pointers from/to another address space can
be consumed by kvm(3) to query for data in kernel space, arrays
are more tricky, especially when their content is copied in userland:
they are part of the copied struct.

Address of array members are only valid in their own address space,
in our case userland, which is (fortunately?) different from kernel space.
This breaks the various kvm_read() calls that query for file descriptor
information. Consequence: lsof(1) cannot print filedescriptor information
(starting from 5.99.14), and silently ignores the errors, as using
the userland fdtab (``dt'' variable) is not valid for kernel.

Fix that by using the ``fd_dt'' member of struct filedes, which stores
the address of the fdtab struct in kernel address space. Took a few hours
to understand what was going on with lsof(1), hmmm.

Luckily, fstat(1) uses the proper model (checked about 5min ago). Why
lsof(1) decided not to log an error on kvm_read() is... a good question.

Bump rev.
----------------------------
revision 1.4
date: 2009-08-12 04:53:06 +0800;  author: apb;  state: Exp;  lines: +78 -17;
Make this build again on netbsd-current.  Bump PKGREVISION to 5.
----------------------------
revision 1.3
date: 2008-05-18 07:53:29 +0800;  author: christos;  state: Exp;  lines: +72 -43;
make this work again.
----------------------------
revision 1.2
date: 2005-01-17 22:54:01 +0800;  author: atatat;  state: dead;  lines: +1 -1;
Upgrade to lsof version 4.74, thereby getting rid of most of the
accumulated patches.  Changes related to or affecting NetBSD (or
pkgsrc) from the 00DIST file:

4.74            January 17, 2005
                ...
                Updated lsof for NetBSD 2.99.10 and tested it on a
                system provided by Andrew Brown <atatat@atatdot.net>.
                ...
                Added BSDI, FreeBSD, NetBSD and OpenBSD support for the
                *effnlink member of the inode structure.  This makes
                the lsof LTnlink test run faster on all modified
                dialects and correctly on OpenBSD.

                Added ptyfs support for NetBSD, using modifications
                provided by Andrew Brown.

                Changed the netbsd Configure stanza to look by default
                for system header files in both /usr/include and
                /usr/src.  (The NETBSD_SYS environment variable can
                still be used to select an alternate for /usr/src.)
                ...
----------------------------
revision 1.1
date: 2004-11-29 06:39:07 +0800;  author: atatat;  state: Exp;
Tweak the sys/bufq.h support (it disappeared again), enhance the
procfs support a little, incorporate the author's patches to deal with
netbsd-current version numbers getting longer, and add ptyfs support.

These patches are adapted from an interim release the author made and
will appear in the next full release.

PKGREVISION bumped to 1 for the ptyfs support.
=============================================================================